requirements.txt
pysimplegui
psycopg2
python-dotenv
pytest